% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number to choose from}

\item{k}{number selected}
}
\value{
number of combinations
}
\description{
combination n choose k
}
\examples{
bin_choose(5,2)
bin_choose(5,1:3)
bin_choose(5,0)
}
