% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin cumlative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
cumulative
}
\description{
binominal cumulative
}
\examples{
bin_cumulative(5, 0.5)
bin_cumulative(7, 0.3)
}
