% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
distribution
}
\description{
binominal distribution
}
\examples{
bin_distribution(5, 0.5)
bin_distribution(7, 0.3)
}
