% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mean
}
\description{
binominal mean
}
\examples{
bin_mean(5, 0.5)
bin_mean(10, 0.3)
}
