% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
mode
}
\description{
binominal mode
}
\examples{
bin_mode(5, 0.5)
bin_mode(10, 0.3)
}
