% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
skewness
}
\description{
binominal skewness
}
\examples{
bin_skewness(5, 0.5)
bin_skewness(10, 0.3)
}
